chore: boxStyles no longer separate for useBox
