+++ lib/components/private/CheckableBase/CheckableBase.tsx
@@ -5,7 +5,7 @@ import React, {
 	type ReactNode,
 } from 'react';
 
-import { elementResetStyles } from '../../../styles';
+import { resetStyles } from '../../../styles';
 import { Box } from '../../Box/Box';
 import { Text } from '../../Text/Text';
 import { textStyles } from '../../Text/textStyles';
@@ -62,7 +62,7 @@ export const CheckableBase = forwardRef<HTMLInputElement, CheckableBaseProps>(
 				position="relative"
 				className={[
 					styles.root,
-					elementResetStyles('button'),
+					resetStyles('button'),
 					className,
 					{ [styles.disabled]: disabled },
 				]}
@@ -82,10 +82,7 @@ export const CheckableBase = forwardRef<HTMLInputElement, CheckableBaseProps>(
 					disabled={disabled}
 					type={inputType}
 					pointerEvents={disabled ? 'none' : void 0}
-					className={[
-						elementResetStyles('button'),
-						styles.nativeInput,
-					]}
+					className={[resetStyles('button'), styles.nativeInput]}
 					onClick={handleClick}
 					onChange={onChange}
 				/>
@@ -94,7 +91,7 @@ export const CheckableBase = forwardRef<HTMLInputElement, CheckableBaseProps>(
 					alignItems="center"
 					justifyContent="center"
 					position="relative"
-					className={[styles.checkable, elementResetStyles('button')]}
+					className={[styles.checkable, resetStyles('button')]}
 				>
 					{children}
 				</Box>
@@ -104,7 +101,7 @@ export const CheckableBase = forwardRef<HTMLInputElement, CheckableBaseProps>(
 					width="full"
 					pointerEvents={disabled ? 'none' : void 0}
 					className={[
-						elementResetStyles('button'),
+						resetStyles('button'),
 						textStyles({ size: '4' }),
 						{
 							[styles.label.disabled]: disabled,
